# test/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/feed/quickCreate/option": {
            "get": {
                "operationId": "Feed_quickCreateOption",
                "summary": "快速添加订阅的配置项",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed/quickCreate": {
            "post": {
                "operationId": "Feed_quickCreate",
                "summary": "快速添加订阅",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuickCreateFeed"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feed"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed/import/option": {
            "get": {
                "operationId": "Feed_importByOpmlOption",
                "summary": "导入 OPML 文件的配置项",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed/import": {
            "post": {
                "operationId": "Feed_importByOpml",
                "summary": "从 OPML 文件导入订阅",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "要上传的文件",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/FileUploadDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed/export": {
            "get": {
                "operationId": "Feed_exportByOpml",
                "summary": "导出订阅为 OPML 文件",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed": {
            "post": {
                "operationId": "Feed_create",
                "summary": "创建记录",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFeed"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feed"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            },
            "put": {
                "operationId": "Feed_update",
                "summary": "更新记录",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFeed"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feed"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            },
            "get": {
                "operationId": "Feed_find",
                "summary": "查找所有记录",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "Query options",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindFeed"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
        "/api/feed/{id}": {
            "delete": {
                "operationId": "Feed_delete",
                "summary": "删除记录",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feed"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            },
            "get": {
                "operationId": "Feed_findOne",
                "summary": "查找记录",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feed"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "feed"
                ]
            }
        },
...
    },
    "info": {
        "title": "RSS Impact server docs",
        "description": "RSS Impact server docs",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        },
        "schemas": {
...
            "QuickCreateFeed": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "URL",
                        "example": "https://blog.cmyr.ltd/atom.xml"
                    },
                    "cron": {
                        "type": "string",
                        "enum": [
                            "EVERY_5_MINUTES",
                            "EVERY_10_MINUTES",
                            "EVERY_20_MINUTES",
                            "EVERY_30_MINUTES",
                            "EVERY_1_HOUR",
                            "EVERY_2_HOURS",
                            "EVERY_3_HOURS",
                            "EVERY_4_HOURS",
                            "EVERY_6_HOURS",
                            "EVERY_8_HOURS",
                            "EVERY_12_HOURS",
                            "EVERY_DAY"
                        ],
                        "title": "Cron",
                        "example": "EVERY_10_MINUTES"
                    },
                    "isEnabled": {
                        "type": "boolean",
                        "title": "是否启用",
                        "example": true
                    },
                    "categoryId": {
                        "type": "number",
                        "title": "分类",
                        "example": 1
                    },
                    "hooks": {
                        "title": "Hook列表",
                        "example": [],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Hook"
                        }
                    }
                },
                "required": [
                    "url",
                    "cron",
                    "isEnabled",
                    "categoryId",
                    "hooks"
                ]
            },
            "FileUploadDto": {
                "type": "object",
                "properties": {
                    "file": {
                        "type": "string",
                        "title": "要上传的文件",
                        "format": "binary"
                    }
                },
                "required": [
                    "file"
                ]
            },
...
            "FindFeed": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Feed"
                        }
                    },
                    "total": {
                        "type": "number"
                    },
                    "lastPage": {
                        "type": "number"
                    },
                    "currentPage": {
                        "type": "number"
                    }
                },
                "required": [
                    "data",
                    "total",
                    "lastPage",
                    "currentPage"
                ]
            },
...
        }
    }
}